fs/fat: Use common read/write/close functions
